native image processing pipeline with 53.9MB image being handled in 69.64 milliseconds. Pure typescript/node,nothing else.
