+++ .changeset/shiny-geese-make.md
@@ -1,0 +1,5 @@
+---
+"@d0paminedriven/fs": minor
+---
+
+updates image handling-> isolated to its own service->extended by fs service
